# fix: Correctly handle document's store updates