--- a/packages/cozy-client/package.json
+++ b/packages/cozy-client/package.json
@@ -20,6 +20,7 @@
     "btoa": "^1.2.1",
     "cozy-stack-client": "^57.0.0",
     "date-fns": "2.29.3",
+    "fast-deep-equal": "^3.1.3",
     "json-stable-stringify": "^1.0.1",
     "lodash": "^4.17.13",
     "microee": "^0.0.6",