@Crash-- 2025-02-28
> This is particularly problematic when there are a lot of queries 

This is why we should add a garbage collector for queries! You've the paper for that :p.